test: 유저 삭제 기능 분리에 따른 테스트 코드 클래스 이동
